inno_vehicle_register/requirements.txt
appdirs==1.4.3
asgiref==3.8.1
attrs==25.2.0
cached-property==1.3.1
certifi==2017.7.27.1
chardet==3.0.4
contourpy==1.1.1
cycler==0.12.1
defusedxml==0.5.0
Django==4.2.7
django-cors-headers==4.3.1
djangorestframework==3.14.0
easyocr==1.7.2
et_xmlfile==2.0.0
filelock==3.16.1
fonttools==4.56.0
fsspec==2025.3.0
idna==2.6
imageio==2.35.1
importlib_resources==6.4.5
isodate==0.6.0
Jinja2==3.1.6
kiwisolver==1.4.7
lazy_loader==0.4
lxml==4.6.0
MarkupSafe==2.1.5
matplotlib==3.7.5
mpmath==1.3.0
# NetSDK @ file:///C:/Users/bilgeework/Desktop/zet/BackEnd/inno_vehicle_register/dahua_integration/NetSDK-2.0.0.1-py3-none-win_amd64.whl#sha256=ed5d8d10b3e7dfde3e1d8457d5979bd2c8c0fe86cedcb18064ffd29f4a6f9d84
networkx==3.1
ninja==1.11.1.3
numpy==1.24.4
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
openpyxl==3.1.2
packaging==24.2
pandas==2.0.3
Pillow==9.5.0
platformdirs==4.3.6
pyclipper==1.3.0.post6
pyparsing==3.1.4
python-bidi==0.6.6
python-dateutil==2.9.0.post0
pytz==2025.1
PyYAML==6.0.2
requests==2.18.4
requests-file==2.1.0
requests-toolbelt==0.8.0
scikit-image==0.21.0
scipy==1.10.1
shapely==2.0.7
six==1.11.0
sqlparse==0.5.3
sympy==1.13.1
tifffile==2023.7.10
torch==2.4.1
torchvision==0.19.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==1.22
websocket-client==1.8.0
zeep==4.2.1
zipp==3.20.2
